+++ store/cachekv/internal/memiterator.go
@@ -19,23 +19,21 @@ type memIterator struct {
 	start     []byte
 	end       []byte
 	ascending bool
-	lastKey   []byte
-	deleted   map[string]struct{}
 	valid     bool
 }
 
-func NewMemIterator(start, end []byte, items *BTree, deleted map[string]struct{}, ascending bool) *memIterator {
+func NewMemIterator(start, end []byte, items MemCache, ascending bool) *memIterator {
 	iter := items.tree.Iter()
 	var valid bool
 	if ascending {
 		if start != nil {
-			valid = iter.Seek(newItem(start, nil))
+			valid = iter.Seek(item{key: start})
 		} else {
 			valid = iter.First()
 		}
 	} else {
 		if end != nil {
-			valid = iter.Seek(newItem(end, nil))
+			valid = iter.Seek(item{key: end})
 			if !valid {
 				valid = iter.Last()
 			} else {
@@ -52,8 +50,6 @@ func NewMemIterator(start, end []byte, items *BTree, deleted map[string]struct{}
 		start:     start,
 		end:       end,
 		ascending: ascending,
-		lastKey:   nil,
-		deleted:   deleted,
 		valid:     valid,
 	}
 
@@ -113,21 +109,7 @@ func (mi *memIterator) Key() []byte {
 }
 
 func (mi *memIterator) Value() []byte {
-	item := mi.iter.Item()
-	key := item.key
-	// We need to handle the case where deleted is modified and includes our current key
-	// We handle this by maintaining a lastKey object in the iterator.
-	// If the current key is the same as the last key (and last key is not nil / the start)
-	// then we are calling value on the same thing as last time.
-	// Therefore we don't check the mi.deleted to see if this key is included in there.
-	if _, ok := mi.deleted[string(key)]; ok {
-		if mi.lastKey == nil || !bytes.Equal(key, mi.lastKey) {
-			// not re-calling on old last key
-			return nil
-		}
-	}
-	mi.lastKey = key
-	return item.value
+	return mi.iter.Item().value
 }
 
 func (mi *memIterator) assertValid() {